re.c: fix recursive call arguments and set CC optionally
When calling the recursive varient of matchpattern, it doesn't
call the function with the matchlength argument. This patch fixes it.

Also, set Makefile CC to gcc when GNU/make can't find the default
CC path. This fixes compilation on other Unix-like platforms where
CC defaults to clang (we previously explicitly set this to GCC).